refactor: bank module handlers
